Cut old filewatcher classes and unwanted Git functionality (pull, revert)
