--- BlockServer/fileIO/base_file_event_handler.py
@@ -1,46 +1,0 @@
-# This file is part of the ISIS IBEX application.
-# Copyright (C) 2012-2016 Science & Technology Facilities Council.
-# All rights reserved.
-#
-# This program is distributed in the hope that it will be useful.
-# This program and the accompanying materials are made available under the
-# terms of the Eclipse Public License v1.0 which accompanies this distribution.
-# EXCEPT AS EXPRESSLY SET FORTH IN THE ECLIPSE PUBLIC LICENSE V1.0, THE PROGRAM
-# AND ACCOMPANYING MATERIALS ARE PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES
-# OR CONDITIONS OF ANY KIND.  See the Eclipse Public License v1.0 for more details.
-#
-# You should have received a copy of the Eclipse Public License v1.0
-# along with this program; if not, you can obtain a copy from
-# https://www.eclipse.org/org/documents/epl-v10.php or
-# http://opensource.org/licenses/eclipse-1.0.php
-
-from watchdog.events import FileSystemEventHandler, FileDeletedEvent, FileMovedEvent
-
-
-class BaseFileEventHandler(FileSystemEventHandler):
-    """ The BaseFileEventHandler class
-        Inherit from this class to provide event handling for different kinds of configuration files.
-    """
-
-    def __init__(self, manager):
-        """ Constructor.
-
-        Args:
-        manager : The File Manager.
-        """
-
-        self._manager = manager
-
-    def on_created(self, event):
-        pass
-
-    def on_modified(self, event):
-        pass
-
-    def on_moved(self, event):
-        pass
-
-    def on_deleted(self, event):
-        pass
-
-
