+++ BlockServer/fileIO/file_event_handler.py
@@ -14,33 +14,41 @@
 # https://www.eclipse.org/org/documents/epl-v10.php or
 # http://opensource.org/licenses/eclipse-1.0.php
 
-from base_file_event_handler import BaseFileEventHandler
+from watchdog.events import FileSystemEventHandler
 from server_common.utilities import print_and_log
 
 
-class UnclassifiedFileEventHandler(BaseFileEventHandler):
-    """ The UnclassifiedFileEventHandler class
-
-    Subclasses the FileSystemEventHandler class from the watchdog module. Handles all events on the filesystem and
-    creates/removes available device screens as necessary.
+class FileEventHandler(FileSystemEventHandler):
+    """ The FileEventHandler class. Provides event handling for events on the file system.
     """
-    def __init__(self, file_manager, ignore_directories):
+
+    def __init__(self, vc):
         """ Constructor.
 
         Args:
-            file_manager (UnclassifiedFileManager): Manages IO with the VC repo
+        manager : The File Manager.
         """
-        super(UnclassifiedFileEventHandler, self).__init__(file_manager)
-        self._ignore_directories = ignore_directories
+
+        self._vc = vc
+
+    def on_created(self, event):
+        pass
+
+    def on_moved(self, event):
+        pass
 
     def on_deleted(self, event):
-        if not event.is_directory and not any(s in event.src_path for s in self._ignore_directories):
+        if not event.is_directory:
             message = "Deleted a file at {0}".format(event.src_path)
             print_and_log(message)
             self._manager.add_and_commit(message)
 
     def on_modified(self, event):
-        if not event.is_directory and not any(s in event.src_path for s in self._ignore_directories):
+        if not event.is_directory:
             message = "A file changed at {0}".format(event.src_path)
             print_and_log(message)
             self._manager.add_and_commit(message, event.src_path)
+
+    def add_and_commit(self, message, path=None):
+        self._vc.add(path)
+        self._vc.commit(message)
